July 16, 2024
Player’s Junk dialog screen for score card
Player junk records are now kept

Next
Player's summary junk screen

+++ app/src/main/java/com/golfpvcc/teamscore_rev4/database/model/PlayerJunkRecord.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
     data class PlayerJunkRecord(
     val mPlayerIdx:Int=0,
     val mHoleNumber:Int=0,
-    val mJunkId:Int=0,
+    val mJunkId:Long=0,
     @PrimaryKey(autoGenerate = true)    // autoGenerate is false
     val mId: Long = 0,
 )
